+++ src/tools/validity_state_checkers.js
@@ -58,6 +58,12 @@ const badInput = check(test_bad_input, element => set_msg(element, 'badInput',
 
 
 function customError(element) {
+  /* prevent infinite loops when the custom validators call setCustomValidity(),
+   * which in turn calls this code again */
+  if (element.__hf_custom_validation_running) {
+    return false;
+  }
+
   /* check, if there are custom validators in the registry, and call
    * them. */
   const custom_validators = CustomValidatorRegistry.get(element);
@@ -65,6 +71,7 @@ function customError(element) {
   var valid = true;
 
   if (cvl) {
+    element.__hf_custom_validation_running = true;
     for (let i = 0; i < cvl; i++) {
       const result = custom_validators[i](element);
       if (result !== undefined && ! result) {
@@ -73,6 +80,7 @@ function customError(element) {
         break;
       }
     }
+    delete(element.__hf_custom_validation_running);
   }
 
   /* check, if there are other validity messages already */